PROJFBPRO-69: Refactor functions into ea_rules_access_functions.py + rename ea_login_functions.py
