+++ src/handle_data.py
@@ -1,14 +1,14 @@
 import json
 
 from src.my_credentials import get_credentials
-from src.obtain_a_JWT_token import get_jwt_token
+from src.ea_login_functions import get_jwt_token
 from src.mdr_event_information_extraction_functions import get_mdr_sys_name
 from src.mdr_event_information_extraction_functions import get_attack_id
-from src.receive_list_of_systems_from_ea import get_all_systems
-from src.extract_ea_system_id import get_ea_system_id_from_list_of_systems
-from src.receive_hardening_variation_key_by_id import get_hardening_variation_key_by_id
-from src.get_hardening_id import get_hardening_id
-from src.get_rule_ids import get_rule_ids
+from src.ea_system_identification_functions import get_all_systems
+from src.ea_system_identification_functions import get_ea_system_id_from_list_of_systems
+from src.ea_rules_access_functions import get_hardening_variation_key_by_id
+from src.ea_rules_access_functions import get_hardening_id
+from src.ea_rules_access_functions import get_rule_ids
 from src.get_the_state_of_the_system import get_the_state
 from src.check_mapping import check_mapping
 from src.update_mitigation_status import update_mitigation_status
