+++ test/TestHardeningVariationKey.py
@@ -1,8 +1,8 @@
 import unittest
 
 from src.my_credentials import get_credentials
-from src.obtain_a_JWT_token import get_jwt_token
-from src.receive_hardening_variation_key_by_id import get_hardening_variation_key_by_id
+from src.ea_login_functions import get_jwt_token
+from src.ea_rules_access_functions import get_hardening_variation_key_by_id
 
 
 class TestHardeningVariationKey(unittest.TestCase):
